requirements.txt
ipython
cgen==2017.1
codepy==2017.2
colorama==0.3.9
Fabric==1.14.0
ipywidgets==7.0.1
Jinja2==2.9.6
matplotlib==2.1.0rc1
mpi4py==2.0.0
notebook==5.1.0
numpy==1.13.3
pandas==0.20.3
scipy==1.0.0rc1
snakemq==1.4
starcluster==0.95.6
voluptuous==0.10.5
PyYAML==3.12
